refactor: one approval decision vocabulary for both hosts
The CLI's approval queue and the extension's request panels each carried their
own copy of the approval decision vocabulary — arm construction, rejection
shape, plan and user-question decisions, and the session bypass request. The
`policy.set` literal for a tool-edit bypass was written out identically in both.

This extracts the vocabulary into `src/shared/session/approvalDecision.ts` and
points both hosts at it. `progressView/frontend/events.ts` was nothing but that
vocabulary and is deleted; `BaseRequestPanel` drops from 239 to 46 lines.

432 lines of duplicated vocabulary deleted against 352 added as the shared
implementation plus ~133 for the CLI's translation into it — net +48. The
deduplication is the point; this repo's record on extract-shared-abstraction is
that it does not pay in LOC.

Two departures from the design doc, both because the tree disagreed with it:
the CLI hook-settles three request kinds (toolEdit, retry, externalInquiry),
not the two the doc's capability record named, so the settle arm would have
been permanently dead for the extension — the shared function owns the four
protocol-answered kinds for the CLI and all seven for the extension, with
overloads keeping the CLI's return type runtime-only. And the CLI's own
`ApprovalDecision` is not subsumed: it is the host-hook bag `subscribeApprovals`
reads for rejection provenance, so it is translated into the shared vocabulary
rather than replaced.
